Move menu item role into own file
