+++ libs/components/src/lib/menu-item/definition.ts
@@ -5,8 +5,6 @@ import styles from './menu-item.scss?inline';
 import { MenuItem } from './menu-item';
 import { MenuItemTemplate as template } from './menu-item.template';
 
-export { MenuItemRole } from './menu-item';
-
 /**
  * @internal
  */
@@ -26,3 +24,4 @@ export const menuItemDefinition = defineVividComponent(
  * @param prefix - the prefix to use for the component name
  */
 export const registerMenuItem = createRegisterFunction(menuItemDefinition);
+export { MenuItemRole } from './menu-item-role';
