+++ libs/components/src/lib/menu-item/menu-item-role.ts
@@ -1,0 +1,24 @@
+export const MenuItemRole = {
+	menuitem: 'menuitem',
+	menuitemcheckbox: 'menuitemcheckbox',
+	menuitemradio: 'menuitemradio',
+	presentation: 'presentation',
+} as const;
+
+/**
+ * The types for menu item roles
+ * @public
+ */
+export type MenuItemRole = typeof MenuItemRole[keyof typeof MenuItemRole];
+
+/**
+ * @internal
+ */
+export const roleForMenuItem: {
+	[value in keyof typeof MenuItemRole]: typeof MenuItemRole[value];
+} = {
+	[MenuItemRole.menuitem]: 'menuitem',
+	[MenuItemRole.menuitemcheckbox]: 'menuitemcheckbox',
+	[MenuItemRole.menuitemradio]: 'menuitemradio',
+	[MenuItemRole.presentation]: 'presentation',
+};
