+++ libs/components/src/lib/menu-item/menu-item.template.ts
@@ -6,11 +6,12 @@ import {
 	when,
 } from '@microsoft/fast-element';
 import { classNames } from '@microsoft/fast-web-utilities';
+import { MenuItemRole } from '@vonage/vivid';
 import { affixIconTemplateFactory } from '../../shared/patterns/affix';
 import { Icon } from '../icon/icon';
 import { menuName } from '../menu/definition';
 import type { VividElementDefinitionContext } from '../../shared/design-system/defineVividComponent';
-import { CheckAppearance, MenuItem, MenuItemRole } from './menu-item';
+import { CheckAppearance, MenuItem } from './menu-item';
 
 const getIndicatorIcon = (x: MenuItem) => {
 	if (x.checkedAppearance === CheckAppearance.TickOnly) {
